Filter environment before mapping the federation rules.
Before the environement dictionary is passed to the RuleProcessor all parameters that
are not inheriting from the basestring type should be filtered out.
Additionally, for better control on what parameters should be passed to the
RuleProcessor, one can configure assertion_prefix parameter in the keystone.conf.
Parameter assertion_prefix value defaults to empty string - ''.
Only environment variables with prefix equal to an assertion_prefix value will be
passed to the RuleProcessor object.

Change-Id: I469107633abfa86232e1dc1f1bc64605cd888dc2
Closes-bug: #1291981